refactor(web): rename memory-edit-source to memory-chat-prompts
The module now builds both bridge drafts (edit and import); the old name is a
leftover of the edit-only stage.
